// src/app/core/models/role.ts
export enum Role {
  All = 'All',
  Admin = 'Admin',
  Founder = 'Founder',
  Rector = 'Rector',
  Principale = 'Principale',
  Student = 'Student',
  Inscription = 'Inscription',
  Parent = 'Parent',
  Staff = 'Staff',
  Teacher = 'Teacher',
  Censor = 'Censor',
  Assistant = 'assistant',
  SG = 'SG',
  Secretary = 'Secretary',
  StudyPrefect = 'StudyPrefect',
  Bursar = 'Bursar',
  General_accountant = 'General_accountant',
  Accountant = 'Accountant',
  staffs = 'Staffs',
  direction = 'Direction',
  founder = 'Founder'

}